refactor(contact): enhance AccountService with contact existence checks + rename contact module to support
