+++ pom.xml
@@ -30,7 +30,7 @@
         <module>auth</module>
         <module>shared</module>
         <module>proposals</module>
-        <module>contact</module>
+        <module>support</module>
         <module>bots</module>
         <module>bookmark</module>
         <module>verification</module>
