+++ support/pom.xml
@@ -10,7 +10,7 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>contact</artifactId>
+    <artifactId>support</artifactId>
 
     <properties>
         <maven.compiler.source>21</maven.compiler.source>
