+++ support/src/main/java/pt/estga/contact/ContactStatus.java
