+++ support/src/main/java/pt/estga/contact/config/ContactModuleAutoConfiguration.java
