+++ user/src/main/java/pt/estga/user/controllers/ContactController.java
@@ -2,28 +2,14 @@ package pt.estga.user.controllers;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import pt.estga.shared.aop.SensitiveOperation;
-import pt.estga.shared.dtos.MessageResponseDto;
-import pt.estga.user.dtos.ContactDto;
-import pt.estga.user.dtos.UserContactDto;
-import pt.estga.user.entities.User;
-import pt.estga.user.mappers.UserContactMapper;
-import pt.estga.user.services.AccountService;
 import pt.estga.user.services.UserContactService;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/v1/account/contacts")
@@ -32,39 +18,7 @@ import java.util.stream.Collectors;
 @PreAuthorize("isAuthenticated()")
 public class ContactController {
 
-    private final AccountService accountService;
     private final UserContactService userContactService;
-    private final UserContactMapper userContactMapper;
-
-    @Operation(summary = "Add a new contact", description = "Adds a new contact to the authenticated user's account.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Contact added successfully",
-                    content = @Content(mediaType = "application/json",
-                            schema = @Schema(implementation = MessageResponseDto.class))),
-            @ApiResponse(responseCode = "400", description = "Invalid contact details provided")
-    })
-    @PostMapping
-    public ResponseEntity<?> addContact(
-            @Parameter(description = "Contact details to be added", required = true)
-            @Valid @RequestBody ContactDto request,
-            @AuthenticationPrincipal User user) {
-        accountService.addContact(user, request.value(), request.type());
-        return ResponseEntity.ok(new MessageResponseDto("A new contact has been added to your account."));
-    }
-
-    @Operation(summary = "Get all contacts", description = "Retrieves a list of all contacts for the authenticated user.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Successfully retrieved list of contacts",
-                    content = @Content(mediaType = "application/json",
-                            schema = @Schema(implementation = UserContactDto.class)))
-    })
-    @GetMapping
-    public ResponseEntity<List<UserContactDto>> getContacts(@AuthenticationPrincipal User user) {
-        List<UserContactDto> contactDtos = accountService.getContacts(user).stream()
-                .map(userContactMapper::toDto)
-                .collect(Collectors.toList());
-        return ResponseEntity.ok(contactDtos);
-    }
 
     @Operation(summary = "Check if a contact exists", description = "Checks if a contact exists by its value.")
     @ApiResponses(value = {
@@ -75,38 +29,5 @@ public class ContactController {
             @Parameter(description = "The value of the contact to check", required = true)
             @RequestParam String value) {
         return ResponseEntity.ok(userContactService.existsByValue(value));
-    }
-
-    @Operation(summary = "Request contact verification", description = "Sends a verification message to the specified contact.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Verification message sent successfully",
-                    content = @Content(mediaType = "application/json",
-                            schema = @Schema(implementation = MessageResponseDto.class))),
-            @ApiResponse(responseCode = "404", description = "Contact not found")
-    })
-    @PostMapping("/{id}/verify")
-    public ResponseEntity<?> requestContactVerification(
-            @Parameter(description = "The ID of the contact to verify", required = true)
-            @PathVariable Long id,
-            @AuthenticationPrincipal User user) {
-        accountService.requestContactVerification(user, id);
-        return ResponseEntity.ok(new MessageResponseDto("A verification message has been sent."));
-    }
-
-    @SensitiveOperation
-    @Operation(summary = "Delete a contact", description = "Deletes a contact from the authenticated user's account.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Contact deleted successfully",
-                    content = @Content(mediaType = "application/json",
-                            schema = @Schema(implementation = MessageResponseDto.class))),
-            @ApiResponse(responseCode = "404", description = "Contact not found")
-    })
-    @DeleteMapping("/{contactId}")
-    public ResponseEntity<?> deleteContact(
-            @Parameter(description = "The ID of the contact to be deleted", required = true)
-            @PathVariable Long contactId,
-            @AuthenticationPrincipal User user) {
-        accountService.deleteContact(user, contactId);
-        return ResponseEntity.ok(new MessageResponseDto("Contact has been deleted from your account."));
     }
 }
