+++ user/src/main/java/pt/estga/user/entities/UserContact.java
@@ -22,7 +22,6 @@ public class UserContact {
     @Enumerated(EnumType.STRING)
     private ContactType type;
 
-    @Column(unique = true)
     private String value;
 
     private boolean isVerified;
