+++ verification/src/main/java/pt/estga/verification/services/ActionCodeService.java
@@ -1,5 +1,6 @@
 package pt.estga.verification.services;
 
+import pt.estga.user.entities.UserContact;
 import pt.estga.verification.entities.ActionCode;
 import pt.estga.verification.enums.ActionCodeType;
 import pt.estga.user.entities.User;
@@ -8,7 +9,7 @@ import java.util.Optional;
 
 public interface ActionCodeService {
 
-    ActionCode createAndSave(User user, ActionCodeType type);
+    ActionCode createAndSave(User user, UserContact userContact, ActionCodeType type);
 
     Optional<ActionCode> findByCode(String code);
